requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
matplotlib>=3.7.0
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
albumentations>=1.3.0
pycocotools>=2.0.6
seaborn>=0.12.0
ipython 